Test subject conversion
